deploy

+++ BergerMsfaApi/Models/DemandGeneration/LeadFollowUpModel.cs
@@ -53,7 +53,7 @@ namespace BergerMsfaApi.Models.DemandGeneration
         //public DropdownDetail ProjectStatusLeadCompleted { get; set; }
         public string ProjectStatusLeadCompletedText { get; set; }
         public string ProjectStatusTotalLossRemarks { get; set; }
-        public int ProjectStatusPartialBusinessPercentage { get; set; }
+        public decimal ProjectStatusPartialBusinessPercentage { get; set; }
         public bool HasSwappingCompetition { get; set; }
         public int? SwappingCompetitionId { get; set; }
         //public DropdownDetail SwappingCompetition { get; set; }
@@ -71,14 +71,14 @@ namespace BergerMsfaApi.Models.DemandGeneration
         public string BrandUsedExteriorBrandName { get; set; }
         public string BrandUsedUnderCoatBrandName { get; set; }
         public string BrandUsedTopCoatBrandName { get; set; }
-        public int ActualPaintJobCompletedInteriorPercentage { get; set; }
-        public int ActualPaintJobCompletedExteriorPercentage { get; set; }
-        public int ActualVolumeSoldInteriorGallon { get; set; }
-        public int ActualVolumeSoldInteriorKg { get; set; }
-        public int ActualVolumeSoldExteriorGallon { get; set; }
-        public int ActualVolumeSoldExteriorKg { get; set; }
-        public int ActualVolumeSoldUnderCoatGallon { get; set; }
-        public int ActualVolumeSoldTopCoatGallon { get; set; }
+        public decimal ActualPaintJobCompletedInteriorPercentage { get; set; }
+        public decimal ActualPaintJobCompletedExteriorPercentage { get; set; }
+        public decimal ActualVolumeSoldInteriorGallon { get; set; }
+        public decimal ActualVolumeSoldInteriorKg { get; set; }
+        public decimal ActualVolumeSoldExteriorGallon { get; set; }
+        public decimal ActualVolumeSoldExteriorKg { get; set; }
+        public decimal ActualVolumeSoldUnderCoatGallon { get; set; }
+        public decimal ActualVolumeSoldTopCoatGallon { get; set; }
         public int BusinessAchievementId { get; set; }
         public LeadBusinessAchievementModel BusinessAchievement { get; set; }
 
@@ -153,7 +153,7 @@ namespace BergerMsfaApi.Models.DemandGeneration
         //public DropdownDetail ProjectStatusLeadCompleted { get; set; }
         public string ProjectStatusLeadCompletedText { get; set; }
         public string ProjectStatusTotalLossRemarks { get; set; }
-        public int ProjectStatusPartialBusinessPercentage { get; set; }
+        public decimal ProjectStatusPartialBusinessPercentage { get; set; }
         public bool HasSwappingCompetition { get; set; }
         public int? SwappingCompetitionId { get; set; }
         //public DropdownDetail SwappingCompetition { get; set; }
@@ -171,14 +171,14 @@ namespace BergerMsfaApi.Models.DemandGeneration
         public string BrandUsedExteriorBrandName { get; set; }
         public string BrandUsedUnderCoatBrandName { get; set; }
         public string BrandUsedTopCoatBrandName { get; set; }
-        public int ActualPaintJobCompletedInteriorPercentage { get; set; }
-        public int ActualPaintJobCompletedExteriorPercentage { get; set; }
-        public int ActualVolumeSoldInteriorGallon { get; set; }
-        public int ActualVolumeSoldInteriorKg { get; set; }
-        public int ActualVolumeSoldExteriorGallon { get; set; }
-        public int ActualVolumeSoldExteriorKg { get; set; }
-        public int ActualVolumeSoldUnderCoatGallon { get; set; }
-        public int ActualVolumeSoldTopCoatGallon { get; set; }
+        public decimal ActualPaintJobCompletedInteriorPercentage { get; set; }
+        public decimal ActualPaintJobCompletedExteriorPercentage { get; set; }
+        public decimal ActualVolumeSoldInteriorGallon { get; set; }
+        public decimal ActualVolumeSoldInteriorKg { get; set; }
+        public decimal ActualVolumeSoldExteriorGallon { get; set; }
+        public decimal ActualVolumeSoldExteriorKg { get; set; }
+        public decimal ActualVolumeSoldUnderCoatGallon { get; set; }
+        public decimal ActualVolumeSoldTopCoatGallon { get; set; }
         //public int BusinessAchievementId { get; set; }
         public SaveLeadBusinessAchievementModel BusinessAchievement { get; set; }
